requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.3.2               # via kombu
argon2-cffi==18.3.0       # via django
asn1crypto==0.24.0        # via cryptography
billiard==3.5.0.4         # via celery
celery==4.2.1             # via django-celery-results
certifi==2018.8.24        # via requests
cffi==1.11.5              # via argon2-cffi, cryptography
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
coreapi==2.3.3            # via django-coreapi, django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
cryptography==2.3.1
django-appconf==1.0.2     # via django-axes, django-compressor
django-axes==4.4.0
django-celery-results==1.0.1
django-compressor==2.2
django-coreapi==2.3.0
django-countries==5.3.2
django-filter==2.0.0
django-ipware==2.1.0
django-oauth-toolkit==1.2.0
django-rest-swagger==2.2.0
django-timezone-field==3.0
django[argon2]==2.1.1
djangorestframework==3.8.2
idna==2.7                 # via cryptography, requests
itypes==1.1.0             # via coreapi
jinja2==2.10              # via coreschema
kombu==4.2.1              # via celery
markupsafe==1.0           # via jinja2
oauthlib==2.1.0           # via django-oauth-toolkit
openapi-codec==1.3.2      # via django-rest-swagger
pip-tools==3.0.0
pycountry==18.5.26
pycparser==2.19           # via cffi
pyopenssl==18.0.0         # via requests
python-dateutil==2.7.3
pytz==2018.5              # via celery, django, django-axes, django-timezone-field
pyuca==1.2
rcssmin==1.0.6            # via django-compressor
requests[security]==2.19.1
rjsmin==1.0.12            # via django-compressor
simplejson==3.16.0        # via django-rest-swagger
six==1.11.0               # via argon2-cffi, cryptography, pip-tools, pyopenssl, python-dateutil
sqlparse==0.2.4
testfixtures==6.3.0
uritemplate==3.0.0        # via coreapi
urllib3==1.23             # via requests
uwsgi==2.0.17.1
vine==1.1.4               # via amqp
wheel==0.32.0
